Undo deletion of this file via fresh upload.
Deletion made in error.  Although this file does not differ between v4.0.2 and v4.0.4, the one here has a 'bugfix' which ensures that the cn_dir part of the path is included in the filenames of the runoff and runoff depth files.   This bugfix still does not appear in the NEMO/r4.0-HEAD, so we may want to submit a ticket, but also keep this version of sbcrnf.F90 for the time being.