Update se-orca025_setup_REF
Reverted to EXP00 naming of run directory for EXP_REF_NOTIDE, which should also fix the unusual placement of certain files in the directory structure.    EXP_REF_NOTIDE still appears in one place in this setup script, because we need to copy the configurable namelists from this directory, as in the github repo structure.  This follows our approach of keeping these namelists in the relevant experiment directory.