Update runscript.slurm
Edited "IF" stmt to ensure run doesn't fall over when nn_diaharm is defined for a run with ln_tide=.false. .
i.e. If runs reaches year defined for tidal harmonic analysis, there is a NEMO error trap for a no-tide run, which we were potentially falling into before this change.