Create runscript.slurm for Anemone
This is a runscript for a 2+6-yr integration of GS1p2_full, from 1976-81, with 3 repeats of 1976 for the spinup of the sea-ice  and harmonic analysis for 1981.      The slurm time limit might need extending or the run restarting, as it should cover 6yr/d but this includes 2 extra years and extra calculations for the tidal analysis.