Update traqsr.F90
I've edited your comment only @jasontempestholt, to reflect that mpp_max is a global max across procs., in case the user thought that nksr would somehow be set to the same value across each proc.    Hope you don't mind me committing directly to the IWD branch, as this is only a comment change.