Update RUNS.md to improve consistency of parameter desc.
Just to be certain, I'd like this reviewing please....

It seems like the 3 main runs in this reference parameter description were no longer consistent with what we've been doing re: implementation of calc_tdiss .     As it stood, there was a suggestion to use the default input wave drag file for GS1p1, as defined in the namelist_ref, but I believe we want the _fac2 version.    For GS1p2, where we do calc_tdiss online, there was waffle about the slightly irrelevant input file, which was not used. 

I've attempted to replace some waffle with hopefully clearer waffle.