Update README.md to remove minor typos
In the examples showing how to run the se-eORCA025_setup script, the input parameter "-r $PWD/cfgs/SE-NEMO" should be "-r $PWD/SE-NEMO".    

The sanity check on this is that the setup script in these examples is executed as "./SE-NEMO/scripts/setup/se-eORCA025_setup", i.e. showing that the user's $PWD at the time of execution is the parent of the directory "SE-NEMO" and not its grandparent.